Upgrade Fuel to 1.3

+++ fuel/core/classes/database/pdo/connection.php
@@ -58,6 +58,7 @@ class Database_PDO_Connection extends \Database_Connection
 			'username'   => null,
 			'password'   => null,
 			'persistent' => false,
+			'compress'	 => true,
 		));
 
 		// Clear the connection parameters for security
@@ -76,6 +77,12 @@ class Database_PDO_Connection extends \Database_Connection
 			$attrs[\PDO::ATTR_PERSISTENT] = true;
 		}
 
+		if (in_array(strtolower($this->_db_type), array('mysql', 'mysqli')) and $compress)
+		{
+			// Use client compression with mysql or mysqli (doesn't work with mysqlnd)
+			$attrs[\PDO::MYSQL_ATTR_COMPRESS] = true;
+		}
+
 		try
 		{
 			// Create a new PDO connection
@@ -83,13 +90,22 @@ class Database_PDO_Connection extends \Database_Connection
 		}
 		catch (\PDOException $e)
 		{
-			throw new \Database_Exception($e->getMessage(), $e->getCode(), $e);
+			$error_code = is_numeric($e->getCode()) ? $e->getCode() : 0;
+			throw new \Database_Exception($e->getMessage(), $error_code, $e);
 		}
 
 		if ( ! empty($this->_config['charset']))
 		{
-			// Set the character set
-			$this->set_charset($this->_config['charset']);
+			// Set Charset for SQL Server connection
+			if (strtolower($this->driver_name()) == 'sqlsrv')
+			{
+				$this->_connection->setAttribute(\PDO::SQLSRV_ATTR_ENCODING, \PDO::SQLSRV_ENCODING_SYSTEM);
+			}
+			else
+			{
+				// Set the character set
+				$this->set_charset($this->_config['charset']);
+			}
 		}
 	}
 
@@ -99,6 +115,15 @@ class Database_PDO_Connection extends \Database_Connection
 		$this->_connection = null;
 
 		return true;
+	}
+
+	/**
+	 * Get the current PDO Driver name
+	 * @return string
+	 */
+	public function driver_name()
+	{
+		return $this->_connection->getAttribute(\PDO::ATTR_DRIVER_NAME);
 	}
 
 	public function set_charset($charset)
@@ -121,30 +146,37 @@ class Database_PDO_Connection extends \Database_Connection
 			$benchmark = \Profiler::start("Database ({$this->_instance})", $sql);
 		}
 
-		try
+		// run the query. if the connection is lost, try 3 times to reconnect
+		$attempts = 3;
+
+		do
 		{
 			try
 			{
 				$result = $this->_connection->query($sql);
+				break;
 			}
 			catch (\Exception $e)
 			{
-				// do a reconnect first and try again, before giving up
-				$this->connect();
-				$result = $this->_connection->query($sql);
-			}
-		}
-		catch (\Exception $e)
-		{
-			if (isset($benchmark))
-			{
-				// This benchmark is worthless
-				\Profiler::delete($benchmark);
-			}
+				if (strpos($e->getMessage(), '2006 MySQL') !== false)
+				{
+					$this->connect();
+				}
+				else
+				{
+					if (isset($benchmark))
+					{
+						// This benchmark is worthless
+						\Profiler::delete($benchmark);
+					}
 
-			// Convert the exception in a database exception
-			throw new \Database_Exception($e->getMessage().' with query: "'.$sql.'"');
+					// Convert the exception in a database exception
+					$error_code = is_numeric($e->getCode()) ? $e->getCode() : 0;
+					throw new \Database_Exception($e->getMessage().' with query: "'.$sql.'"', $error_code, $e);
+				}
+			}
 		}
+		while ($attempts-- > 0);
 
 		if (isset($benchmark))
 		{
@@ -159,18 +191,17 @@ class Database_PDO_Connection extends \Database_Connection
 			// Convert the result into an array, as PDOStatement::rowCount is not reliable
 			if ($as_object === false)
 			{
-				$result->setFetchMode(\PDO::FETCH_ASSOC);
+				$result = $result->fetchAll(\PDO::FETCH_ASSOC);
 			}
 			elseif (is_string($as_object))
 			{
-				$result->setFetchMode(\PDO::FETCH_CLASS, $as_object);
+				$result = $result->fetchAll(\PDO::FETCH_CLASS, $as_object);
 			}
 			else
 			{
-				$result->setFetchMode(\PDO::FETCH_CLASS, 'stdClass');
+				$result = $result->fetchAll(\PDO::FETCH_CLASS, 'stdClass');
 			}
 
-			$result = $result->fetchAll();
 
 			// Return an iterator of results
 			return new \Database_Result_Cached($result, $sql, $as_object);
@@ -206,7 +237,7 @@ class Database_PDO_Connection extends \Database_Connection
 		! is_null($like) and $like = str_replace('%', '.*', $like);
 		foreach ($result as $row)
 		{
-			if ( ! is_null($like) and preg_match($like, $row['Field'])) continue;
+			if ( ! is_null($like) and ! preg_match('#'.$like.'#', $row['Field'])) continue;
 			list($type, $length) = $this->_parse_type($row['Type']);
 
 			$column = $this->datatype($type);
